fix(web): step mess inside request quote button form
