+++ apps/web/app/[locale]/(home)/layout.tsx
@@ -1,8 +1,11 @@
 // import { SceneCanvas } from "~/components/scene-canvas.client";
 
+import { ScrollListener } from "~/components/scroll-listener.client";
+
 export default function HomeLayout({ children }: { children: React.ReactNode }) {
   return (
     <>
+      <ScrollListener />
       {/* <div className="fixed inset-0 pointer-events-none">
         <SceneCanvas />
       </div> */}
