+++ apps/web/app/atoms/show-in-header.ts
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const showInHeaderAtom = atom(false);
