--- apps/web/app/components/request-quote-button.client.tsx
@@ -1,371 +1,0 @@
-"use client";
-
-import {
-  Drawer,
-  DrawerContent,
-  DrawerDescription,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerTitle,
-  DrawerTrigger,
-} from "@synoem/ui/components/drawer";
-import { Button } from "@synoem/ui/components/button";
-import { useRequestQuoteForm } from "~/hooks";
-import { useState } from "react";
-import type { SolarPanel, PumpController } from "@synoem/types";
-import { useIsMobile } from "@synoem/ui/hooks/use-mobile";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@synoem/ui/components/dialog";
-import type { ProductTypeId } from "@synoem/config";
-import type { Locale } from "@synoem/config";
-import { useTranslations } from "next-intl";
-import { ArrowLeft, ArrowRight, Loader2, X } from "lucide-react";
-import { Form } from "@synoem/ui/components/form";
-import { FormStepIndicator } from "./form-step-indicator.client";
-import { basicInquirySchema, productInquirySchema } from "@synoem/schema";
-
-import {
-  TermsField,
-  type FormStep,
-  BasicInquiryForm,
-  ProductInquiryForm,
-  CompanyInquiryForm,
-} from "./inquiry-form";
-import { motion, AnimatePresence } from "motion/react";
-import dynamic from "next/dynamic";
-import { SubmissionConfirmation } from "./submission-confirmation.client";
-import {
-  mobileNavigationOpenAtom,
-  requestQuoteMobileDrawerOpenAtom,
-  scrollLockAtom,
-} from "~/atoms";
-import { useAtom, useSetAtom } from "jotai";
-
-const Turnstile = dynamic(
-  () => import("~/components/cloudflare-turnstile.client").then((mod) => mod.CloudflareTurnstile),
-  {
-    ssr: false,
-  },
-);
-
-type Props = {
-  product?: Pick<SolarPanel | PumpController, "modelName" | "id">;
-  productTypeId?: ProductTypeId;
-  locale: Locale;
-  buttonText?: string;
-  // shimmer?: boolean;
-} & React.ComponentProps<typeof Button>;
-
-export const RequestQuoteButton = ({ product, productTypeId, locale, ...props }: Props) => {
-  const isMobile = useIsMobile();
-
-  const t = useTranslations("RequestQuoteButton");
-
-  const steps: FormStep[] = [
-    {
-      id: "basic",
-      title: t("basicInformation.title"),
-      schema: basicInquirySchema,
-      component: <BasicInquiryForm />,
-    },
-    {
-      id: "product",
-      title: t("productInformation.title"),
-      schema: productInquirySchema,
-      component: <ProductInquiryForm />,
-    },
-    {
-      id: "contact",
-      title: t("companyInformation.title"),
-      schema: basicInquirySchema,
-      component: <CompanyInquiryForm />,
-    },
-  ];
-
-  if (isMobile) {
-    return (
-      <RequestQuoteMobile
-        product={product}
-        productTypeId={productTypeId}
-        locale={locale}
-        steps={steps}
-        {...props}
-      />
-    );
-  }
-
-  return (
-    <RequestQuoteDesktop
-      product={product}
-      productTypeId={productTypeId}
-      locale={locale}
-      steps={steps}
-      {...props}
-    />
-  );
-};
-
-const RequestQuoteMobile = ({
-  product,
-  productTypeId,
-  locale,
-  steps,
-  buttonText,
-  // shimmer = false,
-  ...props
-}: Props & { steps: FormStep[] }) => {
-  const [open, setOpen] = useAtom(requestQuoteMobileDrawerOpenAtom);
-
-  const { form, step, isSubmitting, handlePrevStep, handleNextStep } = useRequestQuoteForm({
-    productTypeId,
-    product,
-    steps,
-  });
-
-  if (!form) {
-    return null;
-  }
-
-  const t = useTranslations("RequestQuoteButton");
-
-  const setScrollLock = useSetAtom(scrollLockAtom);
-  const setMobileNavigationOpen = useSetAtom(mobileNavigationOpenAtom);
-
-  const handleOpenChange = (open: boolean) => {
-    setMobileNavigationOpen(false);
-    setOpen(open);
-
-    setScrollLock(false);
-  };
-
-  const text = buttonText ?? t("requestQuote");
-
-  return (
-    <>
-      <Drawer
-        open={open}
-        onOpenChange={handleOpenChange}
-        disablePreventScroll
-        preventScrollRestoration
-      >
-        <DrawerTrigger asChild>
-          {/** TODO: polish the button */}
-          <Button
-            onClick={(e) => {
-              e.currentTarget.blur();
-            }}
-          >
-            {text}
-          </Button>
-        </DrawerTrigger>
-        <DrawerContent className="flex flex-col h-full max-h-[90vh] pb-0">
-          <DrawerHeader className="text-left shrink-0 border-b">
-            <DrawerTitle className="text-center">{t("requestQuote")}</DrawerTitle>
-            <DrawerDescription>{t("description")}</DrawerDescription>
-          </DrawerHeader>
-          <div className="p-4">
-            <FormStepIndicator
-              style="progress"
-              steps={steps.map((step) => ({ id: step.id, title: step.title ?? "" }))}
-              currentStep={step}
-            />
-          </div>
-          <Form {...form}>
-            <form
-              onSubmit={form.handleSubmit(handleNextStep, (errors) => {
-                console.log("errors", errors);
-                // TODO: Send to sentry
-              })}
-              className="flex flex-col flex-1 min-h-0 overflow-y-auto"
-            >
-              <div className="p-4 flex-1 min-h-0 overflow-y-auto">{steps[step]?.component}</div>
-
-              <DrawerFooter className="pt-2 sticky border-t bottom-0 w-full backdrop-blur-md shrink-0 bg-background/80 safe-area-bottom">
-                <input type="hidden" {...form.register("token")} />
-                <Turnstile />
-                <TermsField name="terms" />
-                <div className="flex flex-col gap-2">
-                  <Button
-                    variant="outline"
-                    className="w-full"
-                    onClick={handlePrevStep}
-                    disabled={isSubmitting}
-                  >
-                    {step === 0 ? <> {t("buttons.clearAll")}</> : <>{t("buttons.back")}</>}
-                  </Button>
-                  <Button type="submit" className="w-full" disabled={isSubmitting}>
-                    {step < steps.length - 1 ? (
-                      <>{t("buttons.next")}</>
-                    ) : (
-                      <>
-                        {isSubmitting ? (
-                          <Loader2 className="h-4 w-4 animate-spin" />
-                        ) : (
-                          t("buttons.submit")
-                        )}
-                      </>
-                    )}
-                  </Button>
-                </div>
-              </DrawerFooter>
-            </form>
-          </Form>
-        </DrawerContent>
-      </Drawer>
-    </>
-  );
-};
-
-const RequestQuoteDesktop = ({
-  product,
-  productTypeId,
-  steps,
-  buttonText,
-  ...props
-}: Props & { steps: FormStep[] }) => {
-  const {
-    form,
-    isSubmitting,
-    isSuccess,
-    handlePrevStep,
-    step,
-    handleNextStep,
-    error,
-    handleReset,
-  } = useRequestQuoteForm({
-    product,
-    productTypeId,
-    steps,
-  });
-
-  if (!form) {
-    console.log("form is not ready");
-    return null;
-  }
-
-  const [open, setOpen] = useState(false);
-
-  const t = useTranslations("RequestQuoteButton");
-
-  const setScrollLock = useSetAtom(scrollLockAtom);
-
-  const handleOpenChange = (open: boolean) => {
-    setOpen(open);
-    if (open) {
-      setScrollLock(true);
-    } else {
-      setScrollLock(false);
-    }
-  };
-
-  const text = buttonText ?? t("requestQuote");
-
-  return (
-    <Dialog open={open} onOpenChange={handleOpenChange}>
-      <DialogTrigger asChild>
-        <Button {...props}>{text}</Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-[1000px] p-0 overflow-hidden">
-        <DialogHeader className="hidden">
-          <DialogTitle className="text-2xl">{t("requestQuote")}</DialogTitle>
-          <DialogDescription className="text-base">{t("description")}</DialogDescription>
-        </DialogHeader>
-        <div className="grid grid-cols-1 md:grid-cols-2 h-[700px]">
-          <div className="bg-card p-8 flex flex-col h-full">
-            <div className="mb-8">
-              <h2 className="text-2xl font-bold tracking-tight">{t("requestQuote")}</h2>
-              <p className="text-muted-foreground mt-2">{t("description")}</p>
-            </div>
-            <FormStepIndicator
-              style="default"
-              orientation="vertical"
-              steps={steps.map((step) => ({ id: step.id, title: step.title ?? "" }))}
-              currentStep={step}
-            />
-            <div className="mt-auto">
-              <div className="rounded-lg bg-secondary/50 p-4">
-                <h4 className="text-sm font-medium leading-none mb-2">{t("support.title")}</h4>
-                <p className="text-sm text-muted-foreground">{t("support.description")}</p>
-                <Button variant="link" className="p-0 h-auto text-sm mt-2">
-                  {t("support.button")}
-                </Button>
-              </div>
-            </div>
-          </div>
-          <Form {...form}>
-            {isSuccess ? (
-              <SubmissionConfirmation
-                title={t("confirmation.title")}
-                message={t("confirmation.message")}
-                onDismiss={handleReset}
-              />
-            ) : (
-              <form
-                onSubmit={form.handleSubmit(handleNextStep, (errors) => {
-                  console.log("errors", errors);
-                  // TODO: Send to sentry
-                })}
-                className="flex flex-col h-full overflow-y-auto"
-              >
-                <AnimatePresence mode="wait">
-                  <motion.div
-                    className="p-8 flex-1 min-h-0 overflow-y-auto"
-                    key={step}
-                    initial={{ opacity: 0 }}
-                    animate={{ opacity: 1 }}
-                    exit={{ opacity: 0 }}
-                    transition={{ duration: 0.3 }}
-                  >
-                    {steps[step]?.component}
-                  </motion.div>
-                </AnimatePresence>
-                {error && (
-                  <div className="mb-4 rounded bg-destructive/10 p-2 text-destructive">{error}</div>
-                )}
-                <div className="flex flex-col gap-4 p-4 pt-2 backdrop-blur-xl bg-background/50">
-                  <input type="hidden" {...form.register("token")} />
-                  <Turnstile />
-                  <TermsField name="terms" />
-                  <div className="flex justify-between border-t pt-4">
-                    <Button variant="outline" onClick={handlePrevStep} disabled={isSubmitting}>
-                      {step === 0 ? (
-                        <>
-                          <X className="mr-2 h-4 w-4" /> {t("buttons.clearAll")}
-                        </>
-                      ) : (
-                        <>
-                          <ArrowLeft className="mr-2 h-4 w-4" /> {t("buttons.back")}
-                        </>
-                      )}
-                    </Button>
-                    <Button type="submit" disabled={isSubmitting}>
-                      {step < steps.length - 1 ? (
-                        <>
-                          {t("buttons.next")} <ArrowRight className="ml-2 h-4 w-4" />
-                        </>
-                      ) : (
-                        <>
-                          {isSubmitting ? (
-                            <Loader2 className="h-4 w-4 animate-spin" />
-                          ) : (
-                            t("buttons.submit")
-                          )}
-                        </>
-                      )}
-                    </Button>
-                  </div>
-                </div>
-              </form>
-            )}
-          </Form>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-};
