+++ apps/web/app/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./use-request-quote-form";
 export * from "./use-product-filters";
 export * from "./use-product-pagination";
 export * from "./use-filtered-products";
+export * from "./use-scroll-lock";
